TimeFormat

+++ Bill3.java
@@ -11,7 +11,7 @@ public class Bill3 {
 	    String name2 = args[1]; // Gets from the user the secound name 
 	    String name3 = args[2]; // Gets from the user the third name 
 		int bill = Integer.parseInt(args[3]); // Get the bill 
-		bill = Math.ceil((double)bill / 3.0);
+		bill = (int) Math.ceil(bill / 3.0);
 		System.err.println("Dear " +name3 +", "+ name2 +", and " + name1 +": pay " + (double)bill +" Shekels each" ); // Print
